feat: score_id on first_places is a bigint

+++ Entities/FirstPlace.cs
@@ -3,7 +3,7 @@ namespace RealistikOsu.Cron.Entities;
 public class FirstPlace
 {
     public int Id { get; init; }
-    public required int ScoreId { get; init; }
+    public required long ScoreId { get; init; }
     public required int UserId { get; init; }
     public required long Score { get; init; }
     public required int MaxCombo { get; init; }
